requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
clize==4.0.3
decorator==4.2.1
docutils==0.14
idna==2.6
jsonschema==2.6.0
od==1.0
pbr==3.1.1
python-cachetclient==0.2.4
requests==2.18.4
sigtools==2.0.1
six==1.11.0
urllib3==1.22
